Fixed the issue where the player can change the name of a park to something that the new park auto-name-gen will pick later. Also Fixed the similar issue for houses, federal parks, and businesses. Fixed some issues with the manage House Owner thing.
Made it so no two houses in a
Country can be named the same rather than just in a town. This makes it
possible to own houses in multiple towns.

Fixed "Transfer LeaderShip" to "Transfer Leadership" in Manage Town menu

Made it so Console can't use Hud Command

Made it so that all messages from the plugin in the hud are prefixed
with [InspiredNations] so that you know where the msg is coming from.