Began work on giving auto-complete to menu options.
Fixed some typographical errors

Added MoneyInBankHigh and MoneyInBankLow to account for rounding errors.
The low value should be used what subtracting and the high value should
be used when adding. This is because the final value is always rounded
down to avoid fractional pennies from interfering with the actual amount
of money. Keep and eye open for more errors in the future.

Converted all the double math in PlayerMethods for economy taxation to
BigDecimal math. Hopefully this will allow for a heavier set of
arithmetic for when countries get really big. The last thing I want is
huge errors.